+++ ExcelWakaTimeAddin/WakaTimeAddIn.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Reflection;
 using System.Threading.Tasks;
-using ExcelWakaTimeAddin.Forms;
 using WakaTime.Forms;
 using Excel = Microsoft.Office.Interop.Excel;
 using WakaTime.Shared.ExtensionUtils;
+using WakaTime.ExtensionUtils;
 
 namespace ExcelWakaTimeAddin
 {
@@ -13,9 +13,9 @@ namespace ExcelWakaTimeAddin
         internal static SettingsForm SettingsForm;
         internal static WakaTime.Shared.ExtensionUtils.WakaTime WakaTime;
 
-        private void WakaTimeAddIn_Startup(object sender, EventArgs e)
+        private async void WakaTimeAddIn_Startup(object sender, EventArgs e)
         {
-            var configuration = new Configuration
+            var metadata = new Metadata
             {
                 EditorName = "excel",
                 PluginName = "excel-wakatime",
@@ -23,19 +23,23 @@ namespace ExcelWakaTimeAddin
                 PluginVersion = Constants.PluginVersion
             };
 
-            WakaTime = new WakaTime.Shared.ExtensionUtils.WakaTime(null, configuration, new Logger());
+            WakaTime = new WakaTime.Shared.ExtensionUtils.WakaTime(metadata, new Logger(Dependencies.GetConfigFilePath()));
 
             WakaTime.Logger.Debug("Initializing in background thread.");
-            Task.Run(() => { InitializeAsync(); }).ContinueWith(t => OnStartupComplete());
+
+            await InitializeAsync();
+
+            // Prompt for api key if not already set
+            if (string.IsNullOrEmpty(WakaTime.Config.GetSetting("api_key")))
+                PromptApiKey();
         }
 
-        private void InitializeAsync()
+        private async Task InitializeAsync()
         {
             try
             {
                 // Settings Form
-                SettingsForm = new SettingsForm(ref WakaTime);
-                SettingsForm.ConfigSaved += SettingsFormOnConfigSaved;                
+                SettingsForm = new SettingsForm(WakaTime.Config, WakaTime.Logger);
 
                 // setup event handlers                
                 Application.WorkbookOpen += ApplicationOnWorkbookOpen;
@@ -43,7 +47,7 @@ namespace ExcelWakaTimeAddin
                 Application.WindowActivate += ApplicationOnWindowActivate;
                 Application.WorkbookActivate += ApplicationOnWorkbookActivate;
 
-                WakaTime.InitializeAsync();
+                await WakaTime.InitializeAsync();
             }
             catch (Exception ex)
             {
@@ -57,7 +61,7 @@ namespace ExcelWakaTimeAddin
         {
             try
             {
-                WakaTime.HandleActivity(wb.FullName, false, string.Empty);
+                WakaTime.HandleActivity(wb.FullName, false, "Microsoft Office");
             }
             catch (Exception ex)
             {
@@ -69,7 +73,7 @@ namespace ExcelWakaTimeAddin
         {
             try
             {
-                WakaTime.HandleActivity(wb.FullName, false, string.Empty);
+                WakaTime.HandleActivity(wb.FullName, false, "Microsoft Office");
             }
             catch (Exception ex)
             {
@@ -81,7 +85,7 @@ namespace ExcelWakaTimeAddin
         {
             try
             {
-                WakaTime.HandleActivity(wb.FullName, true, string.Empty);
+                WakaTime.HandleActivity(wb.FullName, true, "Microsoft Office");
             }
             catch (Exception ex)
             {
@@ -93,24 +97,12 @@ namespace ExcelWakaTimeAddin
         {
             try
             {
-                WakaTime.HandleActivity(wb.FullName, false, string.Empty);
+                WakaTime.HandleActivity(wb.FullName, false, "Microsoft Office");
             }
             catch (Exception ex)
             {
                 WakaTime.Logger.Error("ApplicationOnWorkbookOpen", ex);
             }
-        }        
-
-        private static void OnStartupComplete()
-        {
-            // Prompt for api key if not already set
-            if (string.IsNullOrEmpty(WakaTime.Config.ApiKey))
-                PromptApiKey();
-        }
-
-        private static void SettingsFormOnConfigSaved(object sender, EventArgs eventArgs)
-        {
-            WakaTime.Config.Read();
         }
 
         #endregion
@@ -121,7 +113,7 @@ namespace ExcelWakaTimeAddin
         {
             WakaTime.Logger.Info("Please input your api key into the wakatime window.");
 
-            var form = new ApiKeyForm(ref WakaTime);
+            var form = new ApiKeyForm(WakaTime.Config, WakaTime.Logger);
             form.ShowDialog();
         }